requirements.txt
absl-py==1.4.0
asgiref==3.6.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1670263926556/work
astunparse==1.6.3
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
bcrypt==4.0.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1679481329611/work
contourpy==1.0.7
cycler==0.11.0
debugpy @ file:///D:/bld/debugpy_1680755658559/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
Django==4.2
dlib @ file:///D:/bld/dlib_1679613867288/work
dnspython==2.3.0
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1667317341051/work
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
flatbuffers==23.3.3
fonttools==4.39.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.53.0
h5py==3.8.0
idna==3.4
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1681574084475/work
importlib-resources==5.12.0
imutils==0.5.4
ipykernel @ file:///D:/bld/ipykernel_1679336472934/work
ipython @ file:///D:/bld/ipython_1680185618122/work
itsdangerous==2.1.2
jax==0.4.8
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1669134318875/work
Jinja2==3.1.2
joblib==1.2.0
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1681432441054/work
jupyter_core @ file:///D:/bld/jupyter_core_1678994368923/work
keras==2.12.0
keras-nightly==2.5.0.dev2021032900
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
ml-dtypes==0.1.0
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
numpy @ file:///D:/bld/numpy_1675642778044/work
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opt-einsum==3.3.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1681337016113/work
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.5.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1679871349196/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1677600924538/work
protobuf==4.22.3
psutil @ file:///D:/bld/psutil_1667886042967/work
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1681142969746/work
pymongo==4.3.3
pyparsing==3.0.9
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pywin32==304
pyzmq @ file:///D:/bld/pyzmq_1679317071559/work
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
SQLAlchemy==2.0.9
sqlparse==0.4.3
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==1.1.0
threadpoolctl==3.1.0
tornado @ file:///D:/bld/tornado_1666788767305/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1675110562325/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1678559861143/work
tzdata==2023.3
urllib3==1.26.15
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1673864653149/work
Werkzeug==2.2.3
wincertstore==0.2
wrapt==1.12.1
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1677313463193/work
